change removal release for deprecated driver API calls
The driver_output_term() and driver_send_term() API functions are
deprecated, and the R16 documentation states they will be removed in
Erlang/OTP 17. But since they can't be removed as planned due to internal
use that won't be changed for 17, change the documentation to target
Erlang/OTP 18 instead.

Also fix comments in erts/emulator/beam/io.c to change which release is
targeted for the removal of these functions.